refactor: Move code to separate files
